requirements.txt
discord.py==2.2.3
requests==2.30.0